@@ -21,9 +21,9 @@ zip_safe = False
 packages = find:
 include_package_data = True
 # Add here dependencies of your project (semicolon/line-separated), e.g.
-install_requires = docutils<0.17; dataclasses; configargparse; configparser; elasticsearch>=7.0.0,<7.14.0; statistics; numpy<=1.26.4; pyyaml; requests; redis; python-dateutil>=2.7.3; prometheus_api_client; scipy; openshift==0.11; kubernetes==11; setuptools>=40.3.0; boto3; flent; importlib_metadata; kafka-python; ttp; packaging; distro
+install_requires = docutils; configargparse; configparser; elasticsearch>=7.0.0; statistics; numpy; pyyaml; requests; redis; python-dateutil>=2.7.3; prometheus_api_client; scipy; openshift; kubernetes; setuptools>=40.3.0; boto3; flent; kafka-python; ttp; packaging; distro
 # Require a specific Python version, e.g. Python 2.7 or >= 3.4
-python_requires = >=3.6
+python_requires = >=3.11
 
 [options.extras_require]
 # Add here additional requirements for extra features, to install with:
